fix: amplify 有时候读不到 .env, 因此将写 .env 放到了 prebuild 中
